Split tests into unit and integration tests
