+++ tests/integration/test_agent_registry.py
