+++ tests/integration/test_user_registry.py
